move into review file, remove table
